Differentiate tests according to type
